Restructure filesystem for `feature-flags` package - Rename `build` to `installbinary` - Move code into `src` directory - Move rollup commonjs plugin to dev dependencies -
