+++ packages/feature-flags/bin/flag.js
@@ -1,5 +1,7 @@
-const FlagEditor = require('./FlagEditor');
-const prompts = require('./prompts');
+#!/usr/bin/env node
+
+const FlagEditor = require('../src/FlagEditor');
+const prompts = require('../src/prompts');
 
 const editorInstance = new FlagEditor({ prompts });
