+++ packages/feature-flags/rollup.config.js
@@ -1,7 +1,7 @@
 const commonjs = require('@rollup/plugin-commonjs');
 
 export default {
-    input: 'edit.js',
+    input: 'flag.js',
     output: {
         compact: true,
         dir: 'out',
